+++ samples/CG.Caching.QuickStart/Program.cs
@@ -1,12 +1,14 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using Microsoft.Extensions.Caching.Distributed;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using System;
+using System.Threading.Tasks;
 
 namespace CG.Caching.QuickStart
 {
     class Program
     {
-        static void Main(string[] args)
+        static async Task Main(string[] args)
         {
             var serviceCollection = new ServiceCollection();
 
@@ -14,21 +16,29 @@ namespace CG.Caching.QuickStart
             builder.AddJsonFile("appSettings.json");
             var configuration = builder.Build();
 
-            serviceCollection.AddCaching(
+            serviceCollection.AddDistributedCaching(
                 configuration.GetSection("Services:Caching")
                 );
 
             var serviceProvider = serviceCollection.BuildServiceProvider();
 
-            var cache = serviceProvider.GetRequiredService<ICache>();
+            var cache = serviceProvider.GetRequiredService<IDistributedCache>();
 
             var example1 = new Example()
             {
                 A = "1",
                 B = "2"
             };
-            cache.SetAsync("foo", example1).Wait();
-            var example2 = cache.GetAsync<Example>("foo").Result;
+            
+            await cache.SetAsync(
+                "foo", 
+                example1,
+                new DistributedCacheEntryOptions() { }
+                ).ConfigureAwait(false);
+
+            var example2 = await cache.GetAsync<Example>(
+                "foo"
+                ).ConfigureAwait(false);
 
             // TODO : verify that example2 is equal to example1.
         }