+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # CG.Caching change log
 ---
 
+## 2021.2
+
+* I removed my custom ICache type.
+* I added the extensions for IDistributedCache.
+
+
 ## 2021.1
 
 * I created the package.